Add support for custom headers via -headers flag
It would be nice to add support for custom headers, as some applications
rely on these, such as custom CSRF headers, for each request to pass. I
decided to add a new method to parse/verify flags provided to properly
format the headers in a way we'd need, otherwise we'd need to do this
each time a collector is run, which wouldn't be very efficient. As of
now it only does the header parsing, but it probably can/should be
used for flag validation. One example I noticed is if no arguments
or urls from stdin are provided, the tool just hangs, of which we
could catch with this VerifyFlags method.