Fix bug where URLs with "http" or paths were not being crawled
Hey @hakluke, I noticed the changes you made fixed some aspects, but not all. After some debugging, I noticed that URLs passed from stdin with http prefixes, or any path were not being crawled because that exact url was being passed into colly's `AllowedDomains`. However, that value needs to be a domain name, and not a full URL. My changes will parse the URL and only pass the domain in the `AllowedDomains`